Add afterTransition callback
It is useful when you need to update some state outside of the FSM and
this state should correspond to FSM's state updated in enter()
method(s). You cannot use onTransition, because there, you can learn
what will be the next state, but you don't know what will happen in
that state's enter method.

afterTransition() is called after new state's enter() method so the
information updated there is available.

An example might be an error state, which updates an error message in
its enter() method. To publish the error message outside of the FSM,
you must do it from afterTransition() method.